Move CollectionOf attribute to Cake\ORM\Attribute namespace
