+++ src/ORM/Attribute/CollectionOf.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @since         5.3.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\ORM;
+namespace Cake\ORM\Attribute;
 
 use Attribute;
